Bookservice Admin API - Postgres added

+++ bookservice_admin_api/src/api.rs
@@ -1,15 +1,15 @@
 use serde::{Deserialize, Serialize};
 
-pub type BookId = u64;
+pub type BookId = i32;
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Eq, PartialEq)]
 /// Struct containing book id and title
 pub struct BookTitleAndId {
     pub book_id: BookId,
     pub title: String,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Eq, PartialEq)]
 /// Struct representing details of the book
 pub struct BookDetails {
     pub title: String,
@@ -19,7 +19,7 @@ pub struct BookDetails {
     pub tags: Vec<String>,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Default, Clone, Serialize, Deserialize)]
 /// Struct representing a patch to book details. Allows to specify only a few fields and patch the current details
 pub struct BookDetailsPatch {
     #[serde(skip_serializing_if = "Option::is_none")]
